requirements.txt
pdf2image==1.8.0
tensorflow_serving_api==1.14.0
imutils==0.5.3
pandas==0.25.1
tqdm==4.35.0
scipy==1.3.1
requests==2.22.0
boto3==1.9.228
Werkzeug==0.15.6
imageio==2.5.0
numpy==1.17.2
Flask==1.1.1
botocore==1.12.228
tavern==0.30.3
Shapely==1.6.4.post2
xlrd==1.2.0
sanic==19.6.3
opencv_python==4.1.1.26
Pillow==6.1.0
sacred==0.7.5
scikit-image==0.15.0
scikit_learn==0.21.3
sentry-sdk==0.11.2
gunicorn==19.9.0
aiocontextvars==0.2.2 
whatimage==0.0.3
coloredlogs==10.0
torch==1.2.0                                                             
torchvision==0.4.0 
